refactor: simplify routing structure by removing '/retail' prefix and updating related components for improved navigation
