+++ app/layout.tsx
@@ -149,8 +149,7 @@ export default function RootLayout({
       '@type': 'SearchAction',
       target: {
         '@type': 'EntryPoint',
-        urlTemplate:
-          'https://ingco-service.win/retail?search={search_term_string}',
+        urlTemplate: 'https://ingco-service.win/?search={search_term_string}',
       },
       'query-input': 'required name=search_term_string',
     },
