+++ app/ui/Footer.tsx
@@ -37,12 +37,12 @@ const Footer = () => {
       <ul className="mx-auto grid grid-cols-2 gap-x-6 gap-y-2 md:gap-x-48 lg:flex lg:gap-20 xl:gap-32 xl:pr-[115px] ">
         <li className="flex flex-col lg:order-none">
           <p className="font-medium">Клієнтам</p>
-          <Link className="transition-colors hover:text-white" href="/retail">
+          <Link className="transition-colors hover:text-white" href="/">
             Каталог
             <span className="-mt-2 block text-base">(роздріб)</span>
           </Link>
           <Link className="transition-colors hover:text-white" href="/shop">
-            Кабінет
+            Партнер
           </Link>
           <Link
             className="transition-colors hover:text-white"
